requirements.txt
requests
keyboard
tabulate
colorama
python-dotenv
art